requirements.txt
inflect~=7.0.0
fpdf2
pyfiglet~=1.0.2
emoji~=2.10.1
tabulate~=0.9.0
pillow~=10.2.0
pytest~=7.4.3
numpy~=1.26.4
matplotlib~=3.8.3
requests~=2.31.0